requirements.txt
# Core Data & Numerics
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0
numba>=0.57.0,<1.0.0

# Data Serialization
pyarrow>=12.0.0,<15.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.28.0,<1.0.0

# HTTP Client
aiohttp>=3.8.0,<4.0.0
aiohttp-retry>=2.8.0,<3.0.0

# Configuration & Utilities
pyyaml>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0
tqdm>=4.65.0,<5.0.0

# Visualization
plotly>=5.15.0,<6.0.0